added qn pages.. removed some unused pages

+++ PolyFiveCorp/src/polyfive/entities/EventAttributes.java
@@ -27,7 +27,7 @@ public class EventAttributes {
 		this.eventType = eventType;
 	}
 
-	public double getEventPrice() {
+	public float getEventPrice() {
 		return eventPrice;
 	}
